0 int casting

+++ variadic_functions/0-sum_them_all.c
@@ -17,7 +17,7 @@ int sum_them_all(const unsigned int n, ...)
 
 	va_start(terms, n);
 
-	for (i = 0; i <= n; i++)
+	for (i = 0; i <= (int)n; i++)
 	{
 		res += va_arg(terms, int);
 	}
